add option download maps based on maptypes e.g. "kz", "bhop" etc., update ExtractFile method

+++ BhopMapAutoDownloader/Services/BmdService.cs
@@ -14,9 +14,6 @@ namespace BhopMapAutoDownloader.Services
 {
     public class BmdService
     {
-        private readonly WebClient webclient = new WebClient();
-        private static readonly HttpClientHandler _handler = new HttpClientHandler();
-        private static readonly HttpClient _client = new HttpClient(_handler);
         private static readonly string API_URL = $"https://gamebanana.com/apiv7/Mod/ByCategory?_csvProperties=_idRow,_sName,_aSubmitter,_aFiles,_aGame&_aCategoryRowIds[]=5568&_sOrderBy=_tsDateAdded,DESC&_nPerpage=3";
 
         private readonly DbService _dbservice;
@@ -29,16 +26,20 @@ namespace BhopMapAutoDownloader.Services
             _settings = settings;
             _fileservice = fileservice;
 
-            if (!Directory.Exists(_settings.DownloadPath))
-                Directory.CreateDirectory(_settings.DownloadPath);
-            if (!Directory.Exists(_settings.ExtractPath))
-                Directory.CreateDirectory(_settings.ExtractPath);
+            Directory.CreateDirectory(_settings.DownloadPath);
+            Directory.CreateDirectory(_settings.ExtractPath);
         }
 
         public static async Task<string> GetRecentUploads()
         {
-            using var result = await _client.GetAsync(API_URL);
-            return await result.Content.ReadAsStringAsync();
+            using HttpClient _client = new HttpClient();
+            _client.BaseAddress = new Uri(API_URL);
+            HttpResponseMessage response = await _client.GetAsync(API_URL);
+
+            if (response.IsSuccessStatusCode)
+                return await response.Content.ReadAsStringAsync();
+            
+            return null;
         }
 
         public async Task CheckForNewMaps()
@@ -60,29 +61,34 @@ namespace BhopMapAutoDownloader.Services
                             DownloadLink = items._aFiles[0]._sDownloadUrl
                         };
 
-                        LoggerService.Log($"Found new map: {items._sName} by {items._aSubmitter._sName}");
-                        LoggerService.Log($"Downloading...");
-
-                        _dbservice.AddMap(_toadd);
-                        webclient.DownloadFile(new Uri($"{items._aFiles[0]._sDownloadUrl}"), Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile));
-
-                        LoggerService.Log($"Extracting...");
-                        _fileservice.ExtractFile(items._aFiles[0]._sFile);
-
-                        LoggerService.Log($"Download and extraction of map \"{items._sName}\" completed!", LoggerService.LogType.DONE);
-
-                        if(!_settings.KeepDownloadFiles)
-                            if(File.Exists(Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile)))
-                            {
-                                LoggerService.Log($"Deleting file {items._aFiles[0]._sFile}", LoggerService.LogType.INFO);
-                                File.Delete(Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile));
-                                LoggerService.Log($"File deleted!\n", LoggerService.LogType.DONE);
-                            }
-
-                        if(_settings.EnableFastDlCompression)
+                        if (_settings.MapTypes.Any(m => items._sName.Contains(m, StringComparison.OrdinalIgnoreCase)))
                         {
-                            LoggerService.Log($"Compressing to bz2 for FastDl {items._aFiles[0]._sFile}...", LoggerService.LogType.INFO);
-                            _fileservice.CompressToFastdl(_fileservice.ExtractedFileName);
+                            LoggerService.Log($"Found new map: {items._sName} by {items._aSubmitter._sName}");
+                            LoggerService.Log($"Downloading...");
+
+                            _dbservice.AddMap(_toadd);
+
+                            using WebClient webclient = new WebClient();
+                            webclient.DownloadFile(new Uri($"{items._aFiles[0]._sDownloadUrl}"), Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile));
+
+                            LoggerService.Log($"Extracting...");
+                            _fileservice.ExtractFile(items._aFiles[0]._sFile);
+
+                            LoggerService.Log($"Download and extraction of map \"{items._sName}\" completed!", LoggerService.LogType.DONE);
+
+                            if (!_settings.KeepDownloadFiles)
+                                if (File.Exists(Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile)))
+                                {
+                                    LoggerService.Log($"Deleting file {items._aFiles[0]._sFile}", LoggerService.LogType.INFO);
+                                    File.Delete(Path.Combine(_settings.DownloadPath, items._aFiles[0]._sFile));
+                                    LoggerService.Log($"File deleted!\n", LoggerService.LogType.DONE);
+                                }
+
+                            if (_settings.EnableFastDlCompression)
+                            {
+                                LoggerService.Log($"Compressing to bz2 for FastDl {items._aFiles[0]._sFile}...", LoggerService.LogType.INFO);
+                                _fileservice.CompressToFastdl(_fileservice.ExtractedFileName);
+                            }
                         }
                     }
                 }
